Start client-side
