+++ src/viser/client/package.json
@@ -27,7 +27,7 @@
     "immer": "^10.0.4",
     "its-fine": "^1.2.5",
     "mantine-react-table": "^2.0.0-beta.0",
-    "msgpackr": "^1.8.5",
+    "msgpackr": "^1.10.2",
     "postcss": "^8.4.38",
     "prettier": "^3.0.3",
     "react": "^18.2.0",
